# frontend/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "API Zümm",
    "description" : "SIG apicole de gestion et de suivi de ruchers — contrat REST public. Les identifiants d'API (ex. getZummHoneyActualQuantity) ne se traduisent pas.",
    "license" : {
      "name" : "Propriétaire"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
    "/api/visites/{id}" : {
      "get" : {
        "tags" : [ "visite-controller" ],
        "operationId" : "obtenir",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VisiteReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "visite-controller" ],
        "operationId" : "mettreAJour",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "X-Zumm-Version",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VisiteCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VisiteReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "visite-controller" ],
        "operationId" : "supprimer",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/taches/{id}" : {
      "get" : {
        "tags" : [ "tache-controller" ],
        "operationId" : "obtenir_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TacheReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "tache-controller" ],
        "operationId" : "mettreAJour_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TacheCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TacheReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "tache-controller" ],
        "operationId" : "supprimer_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/sites/{id}" : {
      "get" : {
        "tags" : [ "site-controller" ],
        "operationId" : "obtenir_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SiteReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "site-controller" ],
        "operationId" : "mettreAJour_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SiteCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SiteReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "site-controller" ],
        "operationId" : "supprimer_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/ruches/{id}" : {
      "get" : {
        "tags" : [ "ruche-controller" ],
        "operationId" : "obtenir_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RucheReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "ruche-controller" ],
        "operationId" : "mettreAJour_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RucheCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RucheReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "ruche-controller" ],
        "operationId" : "supprimer_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/plannings/{id}" : {
      "get" : {
        "tags" : [ "planning-controller" ],
        "operationId" : "obtenir_4",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlanningReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "planning-controller" ],
        "operationId" : "mettreAJour_4",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PlanningCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlanningReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "planning-controller" ],
        "operationId" : "supprimer_4",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/materiels/{id}" : {
      "get" : {
        "tags" : [ "materiel-controller" ],
        "operationId" : "obtenir_5",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MaterielReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "materiel-controller" ],
        "operationId" : "mettreAJour_5",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MaterielCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MaterielReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "materiel-controller" ],
        "operationId" : "supprimer_5",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/lots/{id}" : {
      "get" : {
        "tags" : [ "lot-controller" ],
        "operationId" : "obtenir_6",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LotReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "lot-controller" ],
        "operationId" : "mettreAJour_6",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LotCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LotReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "lot-controller" ],
        "operationId" : "supprimer_6",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/fermiers/{id}" : {
      "get" : {
        "tags" : [ "fermier-controller" ],
        "operationId" : "obtenir_7",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FermierReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "fermier-controller" ],
        "operationId" : "mettreAJour_7",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FermierCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FermierReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "fermier-controller" ],
        "operationId" : "supprimer_7",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/fermes/{id}" : {
      "get" : {
        "tags" : [ "ferme-controller" ],
        "operationId" : "obtenir_8",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FermeReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "ferme-controller" ],
        "operationId" : "mettreAJour_8",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FermeCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FermeReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "ferme-controller" ],
        "operationId" : "supprimer_8",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/depenses/{id}" : {
      "put" : {
        "tags" : [ "comptabilite-controller" ],
        "operationId" : "mettreAJour_9",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DepenseCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepenseReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "comptabilite-controller" ],
        "operationId" : "supprimer_9",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/consommables/{id}" : {
      "put" : {
        "tags" : [ "consommable-controller" ],
        "operationId" : "mettreAJour_10",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConsommableCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsommableReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "consommable-controller" ],
        "operationId" : "supprimer_10",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/carnet/gabarits/{id}" : {
      "get" : {
        "tags" : [ "carnet-controller" ],
        "operationId" : "obtenirGabarit",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GabaritReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "carnet-controller" ],
        "operationId" : "mettreAJour_11",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GabaritCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GabaritReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "carnet-controller" ],
        "operationId" : "supprimer_11",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/agents/{id}" : {
      "get" : {
        "tags" : [ "agent-controller" ],
        "operationId" : "obtenir_9",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentReponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "agent-controller" ],
        "operationId" : "mettreAJour_12",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AgentCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "agent-controller" ],
        "operationId" : "supprimer_12",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/photos" : {
      "get" : {
        "tags" : [ "photo-controller" ],
        "operationId" : "lister_8",
        "parameters" : [ {
          "name" : "cible",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "VISITE", "RUCHE", "SITE", "REINE", "RECOLTE", "TRAITEMENT" ]
          }
        }, {
          "name" : "cibleId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PhotoReponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "photo-controller" ],
        "operationId" : "attacher",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PhotoCibleCorps"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PhotoReponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recoltes/{id}" : {
      "get" : {
        "tags" : [ "recolte-controller" ],
        "operationId" : "obtenir_10",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecolteReponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "recolte-controller" ],
        "operationId" : "supprimer_13",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/export/visites" : {
      "get" : {
        "tags" : [ "export-controller" ],
        "operationId" : "visites",
        "parameters" : [ {
          "name" : "format",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "csv"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/visites/{id}/photos/{photoId}" : {
      "delete" : {
        "tags" : [ "visite-controller" ],
        "operationId" : "supprimerPhoto",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "photoId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/varroa/{id}" : {
      "delete" : {
        "tags" : [ "comptage-varroa-controller" ],
        "operationId" : "supprimer_14",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/transports/{id}" : {
      "delete" : {
        "tags" : [ "transport-controller" ],
        "operationId" : "supprimer_15",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/traitements/{id}" : {
      "delete" : {
        "tags" : [ "traitement-controller" ],
        "operationId" : "supprimer_16",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/reines/{id}" : {
      "delete" : {
        "tags" : [ "reine-controller" ],
        "operationId" : "supprimer_17",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/photos/{id}" : {
      "delete" : {
        "tags" : [ "photo-controller" ],
        "operationId" : "supprimer_18",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/nourrissements/{id}" : {
      "delete" : {
        "tags" : [ "nourrissement-controller" ],
        "operationId" : "supprimer_19",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/divisions/{id}" : {
      "delete" : {
        "tags" : [ "division-controller" ],
        "operationId" : "supprimer_20",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/captures/{id}" : {
      "delete" : {
        "tags" : [ "capture-essaim-controller" ],
        "operationId" : "supprimer_21",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TacheReponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "libelle" : {
            "type" : "string"
          },
          "rucheId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "rucheModele" : {
            "type" : "string"
          },
          "agentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "agentNom" : {
            "type" : "string"
          },
          "echeance" : {
            "type" : "string",
            "format" : "date"
          },
          "faite" : {
            "type" : "boolean"
          },
          "priorite" : {
            "type" : "string"
          },
          "categorie" : {
            "type" : "string"
          },
          "origine" : {
            "type" : "string"
          },
          "regleCode" : {
            "type" : "string"
          },
          "creeLe" : {
            "type" : "string",
            "format" : "date-time"
          },
          "majLe" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}